New test: test protocol with real data
